Feat#tests

+++ src/repositories/BaseRepository.ts
@@ -39,16 +39,33 @@ export abstract class BaseRepository<T> {
      
       let updated = await this.findById(id);
       if (!updated) {
-        throw new Error(`Record not found for ID=${id}`);
+
+        const customError: any = new Error(`Record not found for ID=${id}`);
+        customError.message = customError.message;
+        customError.status = 404;
+        customError.code = 'USERNOTFOUND';
+
+        throw customError;
       }
       return updated;
-    } catch (error) {
+    } catch (error: any) {
       throw this.handleError(error, "update");
     }
   }
 
   async delete(id: number): Promise<void> {
     try {
+     
+      let user = await this.findById(id);
+      if (!user) {
+
+        const customError: any = new Error(`Record not found for ID=${id}`);
+        customError.message = customError.message;
+        customError.status = 404;
+        customError.code = 'USERNOTFOUND';
+
+        throw customError;
+      }
       await this.db(this.tableName).where({ id }).delete();
     } catch (error) {
       throw this.handleError(error, "delete");
@@ -59,8 +76,10 @@ export abstract class BaseRepository<T> {
     console.error(`[Repository Error] ${this.tableName}.${action}:`, error);
 
     const customError: any = new Error(`Failed to ${action}`);
-    customError.code = 'REPOSITORY_ERROR'
-    customError.status = 500;
+    customError.code = error.code || "REPOSITORY_ERROR";
+    customError.status = error.status || 500;
+    customError.message = error.message;
+
 
     throw customError;
   }
